Enum networking::Command

pub enum Command {
    Version,
    Verack,
    GetBlocks,
    GetData,
    Block,
    GetHeaders,
    Headers,
    Inv,
    MemPool,
    MerkleBlock,
    CmpctBlock,
    GetBlockTxn,
    BlockTxn,
    SendCmpct,
    NotFound,
    Tx,
    Addr,
    Alert,
    FeeFilter,
    FilterAdd,
    FilterClear,
    FilterLoad,
    GetAddr,
    Ping,
    Pong,
    Reject,
    SendHeaders,
}

A shorthand way of referring to a type of Message. A Command is a single byte, while a Message is about 90 bytes.

Variants

Version

Verack

GetBlocks

GetData

Block

GetHeaders

Headers

Inv

MemPool

MerkleBlock

CmpctBlock

GetBlockTxn

BlockTxn

SendCmpct

NotFound

Tx

Addr

Alert

FeeFilter

FilterAdd

FilterClear

FilterLoad

GetAddr

Ping

Pong

Reject

SendHeaders

Implementations

impl Command

pub fn bytes(&self) -> &[u8; 12]